add SpatialJoinHelpers and move tests to SpatialJoinAlgorithmsTest
